+++ database/factories/ExampleFactory.php
@@ -1,8 +1,9 @@
 <?php
 
+use LaravelEnso\Examples\app\Models\Example;
 use LaravelEnso\Examples\app\Enums\SeniorityEnum;
 
-$factory->define(App\Example::class, function (Faker\Generator $faker) {
+$factory->define(Example::class, function (Faker\Generator $faker) {
     $salary = $faker->numberBetween($min = 1100, $max = 5500);
 
     return [