+++ resources/js/components/enso/bulma/toastr/index.js
@@ -5,31 +5,33 @@ export default {
         const Toastr = Vue.extend(Notification);
 
         Vue.prototype.$toastr = {
-            message(body, title, duration) {
-                this.mount('message', body, title, duration);
-            },
-            primary(body, title, duration) {
-                this.mount('primary', body, title, duration);
-            },
-            success(body, title, duration) {
-                this.mount('success', body, title, duration);
-            },
-            warning(body, title, duration) {
-                this.mount('warning', body, title, duration);
+            error(body, title, duration) {
+                this.toastr('danger', body, title, duration);
             },
             info(body, title, duration) {
-                this.mount('info', body, title, duration);
+                this.toastr('info', body, title, duration);
             },
-            error(body, title, duration) {
-                this.mount('danger', body, title, duration);
+            message(body, title, duration) {
+                this.toastr('message', body, title, duration);
             },
-            mount(type, body, title, duration) {
+            primary(body, title, duration) {
+                this.toastr('primary', body, title, duration);
+            },
+            success(body, title, duration) {
+                this.toastr('success', body, title, duration);
+            },
+            warning(body, title, duration) {
+                this.toastr('warning', body, title, duration);
+            },
+
+            toastr(type, body, title, duration) {
                 if (duration) {
                     options.duration = duration;
                 }
+
                 (new Toastr({
                     propsData: {
-                        duration: duration || options.duration, type, body, title,
+                        ...options, type, body, title,
                     },
                 })).$mount();
             },